add tests, fix low stock props
