+++ react/components/LowStock.tsx
@@ -4,17 +4,16 @@ import styles from '../styles.css'
 
 interface Props {
   text: string
-  threshold: number
+  availableQuantity: number
 }
-// <div className={classNames(styles.productAvailabilityContainer, 'flex pa2')}>
 
-const LowStock: FunctionComponent<Props> = ({ text, threshold }) => {
+const LowStock: FunctionComponent<Props> = ({ text, availableQuantity }) => {
   const [before, after] = text.split('{quantity}')
   return (
     <Fragment>
-      {before && <span className={classNames(styles.lowStockText, 'c-muted-2 t-small')}>{before}</span>}
-      {threshold && <span className={classNames(styles.lowStockHighlight, 'c-muted-2 t-small b mh1')}>{threshold}</span>}
-      {after && <span className={classNames(styles.lowStockText, 'c-muted-2 t-small')}>{after}</span>}
+      {before && <span className={classNames(styles.lowStockText, 'c-muted-2 t-body')}>{before}</span>}
+      {availableQuantity && <span className={classNames(styles.lowStockHighlight, 'c-muted-2 t-body b mh1')}>{availableQuantity}</span>}
+      {after && <span className={classNames(styles.lowStockText, 'c-muted-2 t-body')}>{after}</span>}
     </Fragment>
   )
 }
